Generalize optimization for Endl
Previously, an optimized version of Endl::operator<< was given, to allow
more inlining by GCC. This commit generalizes this optimization to one
applying to all subclasses of Printable, for which the same optimization
should help.

This generalization needs a bit of template magic. Since libstdc++ is
not implemented on AVR, two of the type traits from libstdc++ are
implemented from scratch in this library.